Move schemas to sub directory
