+++ emannotationschemas/schemas/bound_text_tag.py
@@ -1,4 +1,4 @@
-from emannotationschemas.base import BoundSpatialPoint, AnnotationSchema
+from emannotationschemas.schemas.base import BoundSpatialPoint, AnnotationSchema
 import marshmallow as mm
 
 class BoundTagAnnotation(AnnotationSchema):
