+++ emannotationschemas/schemas/categorical_factories.py
@@ -1,4 +1,4 @@
-from emannotationschemas.base import BoundSpatialPoint, AnnotationSchema
+from emannotationschemas.schemas.base import BoundSpatialPoint, AnnotationSchema
 import marshmallow as mm
 
 def BoundCategoricalFactory(allowed_categories,
