+++ emannotationschemas/schemas/cell_type_local.py
@@ -1,4 +1,4 @@
-from emannotationschemas.base import BoundSpatialPoint, AnnotationSchema
+from emannotationschemas.schemas.base import BoundSpatialPoint, AnnotationSchema
 import marshmallow as mm
 from marshmallow.validate import OneOf
