+++ emannotationschemas/schemas/synapse.py
@@ -1,4 +1,4 @@
-from emannotationschemas.base import BoundSpatialPoint, \
+from emannotationschemas.schemas.base import BoundSpatialPoint, \
     SpatialPoint, \
     AnnotationSchema
 from marshmallow import ValidationError
